Fix 5 review issues on divisions directory PR
- Fix grant-matching logic: use funding program intermediate join
  (grant.programId -> funding program -> divisionId -> division)
  instead of checking programId against division keys directly
- Remove unused activeCount prop from DivisionsTable
- Fix "Types" stat to use filteredTypeSummary.length (reflects
  current showAll filter state) instead of typeSummary.length
- Extract duplicated color/label maps into shared division-constants.ts
  importable by both server and client components
- Add aria-pressed to filter badge buttons for accessibility
